Refactor asset system & UI: async loading, new WidgetRenderer
Major overhaul of asset management and UI rendering:
- AssetManager now uses type-safe, thread-safe, templated async loading by default; explicit sync loading available.
- Introduced WidgetRenderer for all widget/UI rendering, replacing NuklearRenderer in editor, gameplay, and asset previews.
- Added in-editor ImGui log console with color-coded log levels.
- Content browser supports C# script creation and .csproj auto-regeneration.
- Improved asset loading feedback and error handling in asset editors.
- Enhanced widget system: better text bounds, asset resolution, and root sizing.
- Modernized codebase, improved thread safety, and updated MochiSharp submodule.

+++ engine/src/ignite/scripting/script_instance.cpp
@@ -200,7 +200,7 @@ namespace ignite
             memcpy(s_FieldValueBuffer, value.data(), copyLen);
         s_FieldValueBuffer[copyLen] = '\0';
 
-        const bool success = m_ScriptHost->SetInstanceFieldValue(m_InstanceId, fieldName, s_FieldValueBuffer, copyLen);
+        const bool success = m_ScriptHost->SetInstanceFieldValue(m_InstanceId, fieldName, s_FieldValueBuffer, (int)copyLen);
         if (success)
         {
             m_ScriptClass->GetInstanceFieldsById(m_InstanceId)->at(fieldName).SetValue<std::string>(std::string(s_FieldValueBuffer));
